Добавила проверки на роли в виде методов checkAccessToComments() & checkAccessToAd()

+++ src/main/java/ru/skypro/homework/controller/CommentsController.java
@@ -4,9 +4,11 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.web.method.annotation.AuthenticationPrincipalArgumentResolver;
 import org.springframework.web.bind.annotation.*;
 import ru.skypro.homework.dto.comments.CommentDTO;
 import ru.skypro.homework.dto.comments.CommentsDTO;
@@ -24,17 +26,16 @@ import ru.skypro.homework.service.interfaces.CommentsService;
 public class CommentsController {
 
     private final CommentsService commentsService;
+
     @Operation(summary = "Получение списка всех комментариев")
-    @PreAuthorize("hasAnyAuthority('ROLE_USER', 'ROLE_ADMIN')")
     @GetMapping("/{id}/comments")
-    public ResponseEntity<CommentsDTO>  getComments(@PathVariable("id") int adId) {
-            CommentsDTO commentsDTO = commentsService.getAllComments(adId);
-            return ResponseEntity.ok(commentsDTO);
+    public ResponseEntity<CommentsDTO> getComments(@PathVariable("id") int adId) {
+        CommentsDTO commentsDTO = commentsService.getAllComments(adId);
+        return ResponseEntity.ok(commentsDTO);
     }
 
     // добавление комментариев
     @Operation(summary = "Добавление нового комментария")
-    @PreAuthorize("hasAnyAuthority('ROLE_USER', 'ROLE_ADMIN')")
     @PostMapping("/{id}/comments")
     public ResponseEntity<CommentDTO> addComment(@PathVariable("id") Integer adId,
                                                  @RequestBody CreateOrUpdateCommentDTO createCommentDto,
@@ -43,22 +44,30 @@ public class CommentsController {
         return ResponseEntity.ok(commentDTO);
     }
 
+
     // удаление комментария по id
     @Operation(summary = "Удаление комментария")
-    @PreAuthorize("hasAuthority('ROLE_ADMIN') or @commentsService.getAuthorByCommentId(#pk).username == authentication.principal.username")
+//    @PreAuthorize("hasAuthority('ROLE_ADMIN') or @commentsService.getAuthorByCommentId(#pk).username == authentication.principal.username")
     @DeleteMapping("/{adId}/comments/{commentId}")
-    public void deleteComment(@PathVariable("adId") int adId , @PathVariable("commentId") int pk ) {
-
-//    public void deleteComment(@PathVariable("commentId") int pk ) {
-      commentsService.deleteComment(adId, pk);
+    public ResponseEntity<?> deleteComment(@PathVariable("adId") int adId, @PathVariable("commentId") int pk,
+                                           Authentication authentication) {
+        if (commentsService.checkAccessToComments(pk, authentication.getName())) {
+            commentsService.deleteComment(adId, pk);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
     // обновление комментария
-        @Operation(summary = "Обновление комментария")
-        @PreAuthorize("hasAuthority('ROLE_ADMIN') or @commentsService.getAuthorByCommentId(#pk).username == authentication.principal.username")
-        @PatchMapping("/{adId}/comments/{commentId}")
-    public ResponseEntity<CommentDTO> updateComment(@PathVariable int adId,@PathVariable("commentId") int pk,
-                                                    @RequestBody CreateOrUpdateCommentDTO updateCommentDTO) {
-              return ResponseEntity.ok(commentsService.updateComment(adId, pk, updateCommentDTO));
+    @Operation(summary = "Обновление комментария")
+//    @PreAuthorize("hasAuthority('ROLE_ADMIN') or @commentsService.getAuthorByCommentId(#pk).username == authentication.principal.username")
+    @PatchMapping("/{adId}/comments/{commentId}")
+    public ResponseEntity<CommentDTO> updateComment(@PathVariable int adId, @PathVariable("commentId") int pk,
+                                                    @RequestBody CreateOrUpdateCommentDTO updateCommentDTO,
+                                                    Authentication authentication) {
+        if (commentsService.checkAccessToComments(pk, authentication.getName())) {
+            return ResponseEntity.ok(commentsService.updateComment(adId, pk, updateCommentDTO));
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 }
